+++ Code/GameObject.java
@@ -5,7 +5,7 @@ import java.awt.Rectangle;
 //It's abstract because it is meant to be extended by other classes
 public abstract class GameObject {
   //protected because it is inherited by other classes
-  protected int x,y,vx,vy;
+  protected double x,y,vx,vy;
   protected String type;
   protected ID id;
   
@@ -34,35 +34,35 @@ public abstract class GameObject {
     this.id = id;
   }
 
-  public int getX() {
+  public double getX() {
     return x;
   }
 
-  public void setX(int x) {
+  public void setX(double x) {
     this.x = x;
   }
 
-  public int getY() {
+  public double getY() {
     return y;
   }
 
-  public void setY(int y) {
+  public void setY(double y) {
     this.y = y;
   }
 
-  public int getVx() {
+  public double getVx() {
     return vx;
   }
 
-  public void setVx(int vx) {
+  public void setVx(double vx) {
     this.vx = vx;
   }
 
-  public int getVy() {
+  public double getVy() {
     return vy;
   }
 
-  public void setVy(int vy) {
+  public void setVy(double vy) {
     this.vy = vy;
   }
